Refactored Alert moving getters and setters to a trait
